Refactoring the move method
