+++ src/main/java/com/hknp/model/dao/DeliveryDAO.java
@@ -126,6 +126,7 @@ public class DeliveryDAO implements IRetrieveEntity<DeliveryEntity, Long>, IModi
    public ArrayList<DeliveryEntity> gets(Integer firstResult, Integer maxResults, String searchKeyword, Integer status, String sortColumnName, String typeSort) {
       EntityManager entityMgr = EntityUtils.getEntityManager();
 
+      EntityUtils utils = new EntityUtils();
       Query query;
       String queryStr;
       Boolean temp;
@@ -135,7 +136,7 @@ public class DeliveryDAO implements IRetrieveEntity<DeliveryEntity, Long>, IModi
                  "or u.userEntity.phoneNumber like :searchKeyword " +
                  "or  u.userEntity.email like :searchKeyword " +
                  "or u.userEntity.gender like :searchKeyword " +
-                 "or u.salary like :searchKeyword " + sortColumn(sortColumnName, typeSort);
+                 "or u.salary like :searchKeyword " + utils.sortColumn(sortColumnName, typeSort);
          query = entityMgr.createQuery(queryStr, DeliveryEntity.class);
       } else {
          temp = status == 1;
@@ -145,7 +146,7 @@ public class DeliveryDAO implements IRetrieveEntity<DeliveryEntity, Long>, IModi
                  "or u.userEntity.phoneNumber like :searchKeyword " +
                  "or  u.userEntity.email like :searchKeyword " +
                  "or u.userEntity.gender like :searchKeyword " +
-                 "or u.salary like :searchKeyword) " + sortColumn(sortColumnName, typeSort);
+                 "or u.salary like :searchKeyword) " + utils.sortColumn(sortColumnName, typeSort);
          query = entityMgr.createQuery(queryStr, DeliveryEntity.class);
          query.setParameter("statusPara", temp);
       }
@@ -165,14 +166,6 @@ public class DeliveryDAO implements IRetrieveEntity<DeliveryEntity, Long>, IModi
          exception.printStackTrace();
       } finally {
          entityMgr.close();
-      }
-      return result;
-   }
-
-   public String sortColumn(String columnName, String typeSort) {
-      String result = "";
-      if (!columnName.equals("")) {
-         result = " ORDER BY u." + columnName + " " + typeSort;
       }
       return result;
    }
